Extract StyleBoxFlat, StyleBoxTexture and StyleBoxLine in their own file
