+++ editor/plugins/canvas_item_editor_plugin.h
@@ -39,12 +39,13 @@ class AcceptDialog;
 class CanvasItemEditorViewport;
 class ConfirmationDialog;
 class EditorData;
-class EditorZoomWidget;
 class EditorSelection;
+class EditorZoomWidget;
 class HScrollBar;
 class HSplitContainer;
 class MenuButton;
 class PanelContainer;
+class StyleBoxTexture;
 class ViewPanner;
 class VScrollBar;
 class VSplitContainer;
