+++ editor/plugins/texture_region_editor_plugin.h
@@ -38,7 +38,7 @@
 #include "scene/3d/sprite_3d.h"
 #include "scene/gui/dialogs.h"
 #include "scene/gui/nine_patch_rect.h"
-#include "scene/resources/style_box.h"
+#include "scene/resources/style_box_texture.h"
 
 class AtlasTexture;
 class OptionButton;
