+++ editor/plugins/visual_shader_editor_plugin.cpp
@@ -59,6 +59,7 @@
 #include "scene/main/window.h"
 #include "scene/resources/curve_texture.h"
 #include "scene/resources/image_texture.h"
+#include "scene/resources/style_box_flat.h"
 #include "scene/resources/visual_shader_nodes.h"
 #include "scene/resources/visual_shader_particle_nodes.h"
 #include "servers/display_server.h"
