+++ scene/gui/color_picker.cpp
@@ -37,6 +37,8 @@
 #include "core/os/os.h"
 #include "scene/gui/color_mode.h"
 #include "scene/resources/image_texture.h"
+#include "scene/resources/style_box_flat.h"
+#include "scene/resources/style_box_texture.h"
 #include "servers/display_server.h"
 #include "thirdparty/misc/ok_color.h"
 #include "thirdparty/misc/ok_color_shader.h"
