+++ scene/gui/graph_edit.cpp
@@ -36,6 +36,7 @@
 #include "scene/gui/box_container.h"
 #include "scene/gui/button.h"
 #include "scene/gui/view_panner.h"
+#include "scene/resources/style_box_flat.h"
 
 constexpr int MINIMAP_OFFSET = 12;
 constexpr int MINIMAP_PADDING = 5;
