+++ tests/scene/test_theme.h
@@ -32,6 +32,7 @@
 #define TEST_THEME_H
 
 #include "scene/resources/image_texture.h"
+#include "scene/resources/style_box_flat.h"
 #include "scene/resources/theme.h"
 #include "tests/test_tools.h"
